Refactor RankTable for seperate Row component
