+++ src/components/RankTable/index.tsx
@@ -1,11 +1,8 @@
-import React, { useState } from "react";
+import React from "react";
 
 import { createStyles, makeStyles } from "@material-ui/core/styles";
 import {
-  Avatar,
-  Box,
   Chip,
-  Collapse,
   Grid,
   IconButton,
   Paper,
@@ -17,18 +14,14 @@ import {
   TableRow,
   Theme,
   Tooltip,
-  Typography,
 } from "@material-ui/core";
-import ClearIcon from "@material-ui/icons/Clear";
 import DirectionsRunIcon from "@material-ui/icons/DirectionsRun";
 import FileCopyIcon from "@material-ui/icons/FileCopy";
-import KeyboardArrowUpIcon from "@material-ui/icons/KeyboardArrowUp";
-import KeyboardArrowDownIcon from "@material-ui/icons/KeyboardArrowDown";
 import VisibilityIcon from "@material-ui/icons/Visibility";
 
-import ProjectDetails from "./ProjectDetails";
-import { Project } from "../types";
+import { Project } from "../../types";
 import red from "@material-ui/core/colors/red";
+import Row from "./Row";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -62,108 +55,6 @@ const useStyles = makeStyles((theme: Theme) =>
     },
   })
 );
-
-interface RowProps {
-  project: Project;
-  i: number;
-  userFavourites: Set<Project["id"]>;
-  toggleFavourite: (id: Project["id"]) => void;
-  swapFavourites: (a: number, b: number) => void;
-  length: number;
-}
-
-const Row: React.FC<RowProps> = ({
-  project,
-  i,
-  userFavourites,
-  toggleFavourite,
-  swapFavourites,
-  length,
-}: RowProps) => {
-  const [open, setOpen] = useState(false);
-
-  const classes = useStyles();
-
-  return (
-    <React.Fragment>
-      <TableRow
-        className={`${classes.root} ${i >= 5 ? classes.lowRank : ""}`}
-        hover
-      >
-        <TableCell align="center">
-          {i === 0 ? (
-            <IconButton disabled>
-              <KeyboardArrowUpIcon />
-            </IconButton>
-          ) : (
-            <Tooltip title="Increase">
-              <IconButton onClick={() => swapFavourites(i, i - 1)}>
-                <KeyboardArrowUpIcon />
-              </IconButton>
-            </Tooltip>
-          )}
-          {i === length - 1 ? (
-            <IconButton disabled>
-              <KeyboardArrowDownIcon />
-            </IconButton>
-          ) : (
-            <Tooltip title="Decrease">
-              <IconButton onClick={() => swapFavourites(i, i + 1)}>
-                <KeyboardArrowDownIcon />
-              </IconButton>
-            </Tooltip>
-          )}
-        </TableCell>
-        <TableCell className={classes.cursor} onClick={() => setOpen(!open)}>
-          <Grid container spacing={2} alignItems="center" wrap="nowrap">
-            <Grid item>
-              <Avatar className={i < 5 ? classes.blue : ""}>
-                {project.id}
-              </Avatar>
-            </Grid>
-            <Grid item>
-              <Typography variant="h6">{project.title}</Typography>
-            </Grid>
-            <Grid item className={classes.right}>
-              {project.allocated ? (
-                <Chip label="Allocated" className={classes.red} />
-              ) : (
-                <Chip label="Unallocated" />
-              )}
-            </Grid>
-          </Grid>
-        </TableCell>
-        <TableCell className={classes.buttonColumn}>
-          <Tooltip title="Show More">
-            <IconButton onClick={() => setOpen(!open)}>
-              {open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
-            </IconButton>
-          </Tooltip>
-          {userFavourites.has(project.id) && (
-            <Tooltip title="Remove">
-              <IconButton onClick={() => toggleFavourite(project.id)}>
-                <ClearIcon />
-              </IconButton>
-            </Tooltip>
-          )}
-        </TableCell>
-      </TableRow>
-      <TableRow>
-        <TableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={6}>
-          <Collapse in={open} timeout="auto" unmountOnExit>
-            <Box margin={1}>
-              <ProjectDetails
-                project={project}
-                isFavourite={userFavourites.has(project.id)}
-                toggleFavourite={() => toggleFavourite(project.id)}
-              />
-            </Box>
-          </Collapse>
-        </TableCell>
-      </TableRow>
-    </React.Fragment>
-  );
-};
 
 interface RankTableProps {
   projects: Project[];
